Added uint macro for unsigned int.

+++ Src/Pipeline/Pipeline.cpp
@@ -1,7 +1,7 @@
 #include "Pipeline.h"
 #include "../Utilities.h"
 
-Pipeline::Pipeline(unsigned int image_buffer_width, unsigned int image_buffer_height)
+Pipeline::Pipeline(uint image_buffer_width, uint image_buffer_height)
 {
 	m_image_buffer_width = image_buffer_width;
 	m_image_buffer_height = image_buffer_height;
@@ -13,7 +13,7 @@ Pipeline::Pipeline(unsigned int image_buffer_width, unsigned int image_buffer_he
 	m_per_sample_processor.push_back(*new PerSampleProcessor(image_buffer_width, image_buffer_height));
 
 	m_image_buffer.resize(m_image_buffer_width * m_image_buffer_height);
-	for (int i = 0; i < m_image_buffer_width * m_image_buffer_height; i++)
+	for (uint i = 0; i < m_image_buffer_width * m_image_buffer_height; i++)
 	{
 		m_image_buffer[i] = new float[3];
 		m_image_buffer[i][0] = 0;
@@ -22,14 +22,14 @@ Pipeline::Pipeline(unsigned int image_buffer_width, unsigned int image_buffer_he
 	}
 }
 
-void Pipeline::renderObject(buffer<float>& VBO, buffer<unsigned int>& VAO)
+void Pipeline::renderObject(buffer<float>& VBO, buffer<uint>& VAO)
 {
 	buffer<float> processed_VBO;
-	buffer<unsigned int> processed_VAO;
+	buffer<uint> processed_VAO;
 
 	processVertices(processed_VBO, processed_VAO, VBO, VAO);
 
-	for (int i = 0; i < processed_VAO.size(); i += 3)
+	for (uint i = 0; i < processed_VAO.size(); i += 3)
 		renderTriangle(processed_VBO, &processed_VAO[i]);
 	saveEXR(m_image_buffer, "test.exr", m_image_buffer_width, m_image_buffer_height);
 }
@@ -37,8 +37,8 @@ void Pipeline::renderObject(buffer<float>& VBO, buffer<unsigned int>& VAO)
 void Pipeline::setPixel(float* pixel_data)
 {
 
-	int x = floor(pixel_data[0]);
-	int y = m_image_buffer_height - 1 - floor(pixel_data[1]);
+	uint x = uint(floor(pixel_data[0]));
+	uint y = uint(m_image_buffer_height - 1 - floor(pixel_data[1]));
 	if (x >= m_image_buffer_width || x < 0 || y >= m_image_buffer_height || y < 0)
 		return;
 	m_image_buffer[x + y * m_image_buffer_width][0] = pixel_data[2];
@@ -46,7 +46,7 @@ void Pipeline::setPixel(float* pixel_data)
 	m_image_buffer[x + y * m_image_buffer_width][2] = pixel_data[4];
 }
 
-void Pipeline::processVertices(buffer<float>& output_VBO, buffer<unsigned int>& output_VAO, buffer<float>& input_VBO, buffer<unsigned int>& input_VAO)
+void Pipeline::processVertices(buffer<float>& output_VBO, buffer<uint>& output_VAO, buffer<float>& input_VBO, buffer<uint>& input_VAO)
 {
 	buffer<float> perspective_VBO;
 	m_vertex_shader[0].processVertices(perspective_VBO, input_VBO);
@@ -62,7 +62,7 @@ void Pipeline::processTriangle(buffer<float>& output_fragments, float* vertex_0,
 	m_fragment_shader[0].shadeFragments(output_fragments, fragments);
 }
 
-void Pipeline::renderTriangle(buffer<float>& input_VBO, unsigned int* input_VAO)
+void Pipeline::renderTriangle(buffer<float>& input_VBO, uint* input_VAO)
 {
 	buffer<float> shaded_fragments;
 	buffer<float> post_processed_fragments;
@@ -73,6 +73,6 @@ void Pipeline::renderTriangle(buffer<float>& input_VBO, unsigned int* input_VAO)
 		&input_VBO[m_vertex_size * input_VAO[2]]
 	);
 	m_per_sample_processor[0].postProcessFragments(post_processed_fragments, shaded_fragments);
-	for (int j = 0; j < post_processed_fragments.size(); j += 5)
+	for (uint j = 0; j < post_processed_fragments.size(); j += 5)
 		setPixel(&post_processed_fragments[j]);
 }
